requirements.txt
certifi==2022.6.15
click==8.1.3
influxdb-client==1.31.0
paho-mqtt==1.6.1
python-dateutil==2.8.2
Rx==3.2.0
six==1.16.0
urllib3==1.26.11
